reconstruct the project
