+++ DecryptLogin/modules/core/baidutieba.py
@@ -6,13 +6,13 @@ Author:
 微信公众号:
     Charles的皮卡丘
 更新日期:
-    2020-10-29
+    2022-03-09
 '''
 import os
 import time
 import json
 import requests
-from ..utils.misc import *
+from ..utils import saveImage, showImage, removeImage
 
 
 '''PC端登录百度贴吧'''
