+++ DecryptLogin/modules/core/douyu.py
@@ -6,7 +6,7 @@ Author:
 微信公众号:
     Charles的皮卡丘
 更新日期:
-    2020-10-29
+    2022-03-09
 '''
 import os
 import re
@@ -14,7 +14,7 @@ import time
 import json
 import qrcode
 import requests
-from ..utils.misc import *
+from ..utils import removeImage, saveImage, showImage
 
 
 '''PC端登录斗鱼直播'''
