+++ DecryptLogin/modules/utils/__init__.py
@@ -3,7 +3,7 @@ from .misc import showImage, removeImage, saveImage
 from .cookies import saveSessionCookies, loadSessionCookies
 
 
-'''define import all'''
+'''all'''
 __all__ = [
     'showImage', 'removeImage', 'saveImage',
     'saveSessionCookies', 'loadSessionCookies',
